* A number of changes to improve performance, flexibility, etc... * In-progress development of model point filters which process a set of coordinates one at a time * Disk caching model grid * Other tweaks

+++ trunk/jdem846/source/main/src/us/wthr/jdem846/model/ModelGrid.java
@@ -17,7 +17,7 @@ public class ModelGrid extends ModelPointGrid
 	{
 		super(north, south, east, west, latitudeResolution, longitudeResolution, minimum, maximum);
 
-		grid = new ModelPoint[gridLength];
+		grid = new ModelPoint[(int)gridLength];
 		
 		log.info("Model Grid Width/Height: " + width + "/" + height + ", Length: " + gridLength);
